requirements.txt
python-dotenv
flask
openai
langchain
langchain-openai
azure-search-documents
azure-identity